Move Role constants class to constants package.
